Fix incomplete save/restore of member variable
The method "void  CBigComplexLinProb::Periodicity(int i, int j)" in file
cspars.h stores the value of the member variable *bdw* in a temporary, but
(unlike the other member functions) does not restore it to its old value.

This fix restores the likely intent and restores the old value after the
computation is done.