Change CMake library setup.
Instead of having a libfemm object library that is included in the
fmesher/*solver/*pproc libraries, it is now a normal library that is linked
against.

Reason: With the previous setup I expect problems when one wants to link to
e.g. both fsolver and hsolver, because the libfemm symbols are duplicated.
With the new setup, this should not be an issue.