Make FemmReader work with plain FemmProblem pointers.
FemmReader stored a shared_ptr<FemmProblem> until now, which is a safe choice,
but forces any FemmProblem data objects to be a shared_ptr (and not, e.g. a unique_ptr).
Also, the shared_ptr implies shared ownership, when in truth the FemmReader is only an
intermediate user of the object.