fmesher: code simplification
Make ptlist in doPeriodicBCTriangulation a vector of CCommonpoint instead of a
vector of unique_ptrs.  In contrast to other places where I changed to vectors
of unique_ptrs, in this case we don't need the extra flexibility of a pointer.
A plain old data object is therefore better suited here.