fmesher: clear tags in cnt flags in the same places as in femm42.
femm42 stores notation data in the IsSelected field and clears it after meshing
using the UnselectAll() function.  xfemm changed the data type of IsSelected
from int to bool and therefore introduced a new field 'cnt' for this purpose.
This commit makes sure that cnt data is cleared between calls to triangle, as
it would have been cleared in femm42.