Change string literal parsing according to FEMM semantics.
FEMM allows the string delimiter to be part of the string in unescaped form
(e.g. "a string literal with the letter "a" in it").  OTOH, FEMM does not use
significant characters on the same line after a string literal, so consuming
the whole line and searching for the last delimiter works.