CMaterialProp: adhere more closely to femm42.
The login introduces in commit 1475:f19e4597a857 to process th bhpoints of a
CMMaterialProp after reading it from disk was slightly different that the one
in femm/FemmviewDoc.cpp.

Since the test failure for femmcli_antiperiodicBC_flux.lua seems to be related
to that code, I'm eliminating this difference.