Correctly outputs problem precision
All 3 of the solver programs incorrectly output an empty string of zeros as the problem precision. Tested on windows (msvc) and wsl-ubuntu (gcc) both have the same issue; "Precision: 0.000000" for some reason the to_string never seems able to work for this so it was added as a std::cout statement just afterwards which seems to fix it.